+++ lib/rumale/linear_model/linear_regression.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
-require 'rumale/linear_model/base_linear_model'
+require 'rumale/linear_model/base_sgd'
 require 'rumale/base/regressor'
 
 module Rumale
   module LinearModel
     # LinearRegression is a class that implements ordinary least square linear regression
-    # with mini-batch stochastic gradient descent optimization or singular value decomposition.
+    # with stochastic gradient descent (SGD) optimization or singular value decomposition (SVD).
     #
     # @example
     #   estimator =
-    #     Rumale::LinearModel::LinearRegression.new(max_iter: 1000, batch_size: 20, random_seed: 1)
+    #     Rumale::LinearModel::LinearRegression.new(max_iter: 500, batch_size: 20, random_seed: 1)
     #   estimator.fit(training_samples, traininig_values)
     #   results = estimator.predict(testing_samples)
     #
@@ -19,7 +19,10 @@ module Rumale
     #   estimator = Rumale::LinearModel::LinearRegression.new(solver: 'svd')
     #   estimator.fit(training_samples, traininig_values)
     #   results = estimator.predict(testing_samples)
-    class LinearRegression < BaseLinearModel
+    #
+    # *Reference*
+    # - L. Bottou, "Large-Scale Machine Learning with Stochastic Gradient Descent," Proc. COMPSTAT'10, pp. 177--186, 2010.
+    class LinearRegression < BaseSGD
       include Base::Regressor
 
       # Return the weight vector.
@@ -36,34 +39,57 @@ module Rumale
 
       # Create a new ordinary least square linear regressor.
       #
+      # @param learning_rate [Float] The initial value of learning rate.
+      #   The learning rate decreases as the iteration proceeds according to the equation: learning_rate / (1 + decay * t).
+      #   If solver = 'svd', this parameter is ignored.
+      # @param decay [Float] The smoothing parameter for decreasing learning rate as the iteration proceeds.
+      #   If nil is given, the decay sets to 'learning_rate'.
+      #   If solver = 'svd', this parameter is ignored.
+      # @param momentum [Float] The momentum factor.
+      #   If solver = 'svd', this parameter is ignored.
       # @param fit_bias [Boolean] The flag indicating whether to fit the bias term.
       # @param bias_scale [Float] The scale of the bias term.
-      # @param max_iter [Integer] The maximum number of iterations.
+      # @param max_iter [Integer] The maximum number of epochs that indicates
+      #   how many times the whole data is given to the training process.
       #   If solver = 'svd', this parameter is ignored.
       # @param batch_size [Integer] The size of the mini batches.
       #   If solver = 'svd', this parameter is ignored.
-      # @param optimizer [Optimizer] The optimizer to calculate adaptive learning rate.
-      #   If nil is given, Nadam is used.
+      # @param tol [Float] The tolerance of loss for terminating optimization.
       #   If solver = 'svd', this parameter is ignored.
-      # @param solver [String] The algorithm to calculate weights. ('sgd' or 'svd').
+      # @param solver [String] The algorithm to calculate weights. ('auto', 'sgd' or 'svd').
+      #   'auto' chooses the 'svd' solver if Numo::Linalg is loaded. Otherwise, it chooses the 'sgd' solver.
       #   'sgd' uses the stochastic gradient descent optimization.
       #   'svd' performs singular value decomposition of samples.
       # @param n_jobs [Integer] The number of jobs for running the fit method in parallel.
       #   If nil is given, the method does not execute in parallel.
       #   If zero or less is given, it becomes equal to the number of processors.
       #   This parameter is ignored if the Parallel gem is not loaded.
+      # @param verbose [Boolean] The flag indicating whether to output loss during iteration.
+      #   If solver = 'svd', this parameter is ignored.
       # @param random_seed [Integer] The seed value using to initialize the random generator.
-      def initialize(fit_bias: false, bias_scale: 1.0, max_iter: 1000, batch_size: 10, optimizer: nil,
-                     solver: 'sgd', n_jobs: nil, random_seed: nil)
-        check_params_numeric(bias_scale: bias_scale, max_iter: max_iter, batch_size: batch_size)
-        check_params_boolean(fit_bias: fit_bias)
+      def initialize(learning_rate: 0.01, decay: nil, momentum: 0.9,
+                     fit_bias: true, bias_scale: 1.0, max_iter: 200, batch_size: 50, tol: 1e-4,
+                     solver: 'auto',
+                     n_jobs: nil, verbose: false, random_seed: nil)
+        check_params_numeric(learning_rate: learning_rate, momentum: momentum,
+                             bias_scale: bias_scale, max_iter: max_iter, batch_size: batch_size)
+        check_params_boolean(fit_bias: fit_bias, verbose: verbose)
         check_params_string(solver: solver)
-        check_params_numeric_or_nil(n_jobs: n_jobs, random_seed: random_seed)
-        check_params_positive(max_iter: max_iter, batch_size: batch_size)
-        keywd_args = method(:initialize).parameters.map { |_t, arg| [arg, binding.local_variable_get(arg)] }.to_h.merge(reg_param: 0.0)
-        keywd_args.delete(:solver)
-        super(keywd_args)
-        @params[:solver] = solver != 'svd' ? 'sgd' : 'svd'
+        check_params_numeric_or_nil(decay: decay, n_jobs: n_jobs, random_seed: random_seed)
+        check_params_positive(learning_rate: learning_rate, max_iter: max_iter, batch_size: batch_size)
+        super()
+        @params.merge!(method(:initialize).parameters.map { |_t, arg| [arg, binding.local_variable_get(arg)] }.to_h)
+        @params[:solver] = if solver == 'auto'
+                             load_linalg? ? 'svd' : 'sgd'
+                           else
+                             solver != 'svd' ? 'sgd' : 'svd'
+                           end
+        @params[:decay] ||= @params[:learning_rate]
+        @params[:random_seed] ||= srand
+        @rng = Random.new(@params[:random_seed])
+        @loss_func = LinearModel::Loss::MeanSquaredError.new
+        @weight_vec = nil
+        @bias_term = nil
       end
 
       # Fit the model with given training data.
@@ -94,33 +120,11 @@ module Rumale
         x.dot(@weight_vec.transpose) + @bias_term
       end
 
-      # Dump marshal data.
-      # @return [Hash] The marshal data about LinearRegression.
-      def marshal_dump
-        { params: @params,
-          weight_vec: @weight_vec,
-          bias_term: @bias_term,
-          rng: @rng }
-      end
-
-      # Load marshal data.
-      # @return [nil]
-      def marshal_load(obj)
-        @params = obj[:params]
-        @weight_vec = obj[:weight_vec]
-        @bias_term = obj[:bias_term]
-        @rng = obj[:rng]
-        nil
-      end
-
       private
 
       def fit_svd(x, y)
-        samples = @params[:fit_bias] ? expand_feature(x) : x
-
-        s, u, vt = Xumo::Linalg.svd(samples, driver: 'sdd', job: 'S')
-        d = (s / s**2).diag
-        w = vt.transpose.dot(d).dot(u.transpose).dot(y)
+        x = expand_feature(x) if fit_bias?
+        w = Numo::Linalg.pinv(x, driver: 'svd').dot(y)
 
         is_single_target_vals = y.shape[1].nil?
         if @params[:fit_bias]
@@ -150,8 +154,14 @@ module Rumale
         end
       end
 
-      def calc_loss_gradient(x, y, weight)
-        2.0 * (x.dot(weight) - y)
+      def fit_bias?
+        @params[:fit_bias] == true
+      end
+
+      def load_linalg?
+        return false if defined?(Numo::Linalg).nil?
+        return false if Numo::Linalg::VERSION < '0.1.4'
+        true
       end
     end
   end